Question Generation/requirements.txt
litellm>=1.0.0
langfuse>=2.0.0
python-dotenv>=1.0.0
pydantic>=2.0.0
pandas>=2.0.0

